% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEL0550.R
\name{FEL0550}
\alias{FEL0550}
\title{Transmission from Thorlabs longpass filter number FEL0550 given wavelength}
\usage{
FEL0550(wavelength_nm)
}
\arguments{
\item{wavelength_nm}{incident wavelength in nanometers}
}
\value{
Returns the fractional transmission coefficient of Thorlabs longpass filter number FEL0550
}
\description{
Transmission from Thorlabs longpass filter number FEL0550 given wavelength
}
\examples{
FEL0550(570)
}
