% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEL0600.R
\name{FEL0600}
\alias{FEL0600}
\title{Transmission from Thorlabs longpass filter number FEL0600 given wavelength}
\usage{
FEL0600(wavelength_nm)
}
\arguments{
\item{wavelength_nm}{incident wavelength in nanometers}
}
\value{
Returns the fractional transmission coefficient of Thorlabs longpass filter number FEL0600
}
\description{
Transmission from Thorlabs longpass filter number FEL0600 given wavelength
}
\examples{
FEL0600(620)
}
