% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FES0550.R
\name{FES0550}
\alias{FES0550}
\title{Transmission from Thorlabs shortpass filter number FES0550 given wavelength}
\usage{
FES0550(wavelength_nm)
}
\arguments{
\item{wavelength_nm}{incident wavelength in nanometers}
}
\value{
Returns the fractional transmission coefficient of Thorlabs shortpass filter number FES0550
}
\description{
Transmission from Thorlabs shortpass filter number FES0550 given wavelength
}
\examples{
FES0550(570)
}
