% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_cursor.R
\name{mdb_cursor}
\alias{mdb_cursor}
\title{Use mdb transactions}
\description{
Cursors are required for some advanced queries on an mdb database,
when the basic set of functions in \code{\link{mdb_txn}} is not
sufficient.
}
\details{
Cursors must be created from within a transaction (which in turn
are created from an environment).
}
