% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_txn.R
\name{mdb_txn}
\alias{mdb_txn}
\title{Use mdb transactions}
\description{
Transactions are required for every mdb operation.  Even when
using the convenience functions in \code{\link{mdb_env}}
(\code{get}, etc), a transaction is created and committed each
time.  Within a transaction, either everything happens or nothing
happens, and everything gets a single consistent view of the
database.
}
\details{
There can be many read transactions per environment, but only one
write transactions.  Because R is single-threaded, that means that
you can only simultaneously write from an mdb environment from a
single object - any further attempts to open write transactions it
would block forever while waiting for a lock that can't be
released because there is only one thread!
}
