% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storr.R
\name{storr_thor}
\alias{storr_thor}
\alias{driver_thor}
\title{Thor driver for storr}
\usage{
storr_thor(
  env,
  prefix = "",
  hash_algorithm = NULL,
  default_namespace = "objects"
)

driver_thor(env, prefix = "", hash_algorithm = NULL)
}
\arguments{
\item{env}{A thor environment}

\item{prefix}{An optional prefix.  If given, use a \code{:} as the
last character for nice looking keys (e.g., \code{storr:} will
generate keys like \code{storr:keys:namespace:name}.  If not
given then we assume that storr is the only user of this
database and if \code{destroy} is called it will delete the
entire database.}

\item{hash_algorithm}{Optional hash algorithm to use.  Defaults to
md5, or whatever the existing algorithm is if the database has
been opened.  You cannot mix algorithms.}

\item{default_namespace}{The default namespace to store objects
in.  Defaults to \code{objects}, as does other storr drivers.}
}
\description{
Storr driver for thor.  This allows thor to be used as a storage
backend with the storr package and presents a higher level content
addressable key/value store suitable for storing R objects.
}
